Extract exec action
